pyproject.toml
[project]
name = "agntrick-whatsapp"
version = "0.3.5"
description = "WhatsApp integration for Agntrick agents"
readme = "README.md"
requires-python = ">=3.12,<3.15"
authors = [
    {name = "Jean Silva", email = "me@jeancsil.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["ai", "agents", "whatsapp", "messaging", "llm"]

dependencies = [
    "agntrick>=0.2.9",
    "neonize",
    "ffmpeg-python>=0.2.0",
    "pydantic>=2.12.5",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "dateparser>=1.3.0",
    "croniter>=6.0.0",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.urls]
Homepage = "https://github.com/jeancsil/agntrick-whatsapp"
Documentation = "https://github.com/jeancsil/agntrick-whatsapp#readme"
Repository = "https://github.com/jeancsil/agntrick-whatsapp"
Issues = "https://github.com/jeancsil/agntrick-whatsapp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agntrick_whatsapp"]

[tool.pytest.ini_options]
addopts = "-ra -q --import-mode=importlib"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
    "types-croniter>=6.0.0.20250809",
    "types-python-dateutil>=2.9.0.20260305",
]
